Expose "fileopt" option
fileopt was previously hardcoded to “overwrite”, so whenever a graph
was initialized, (`graph.init();`) the data was overwritten. This is
the perfect option for development: when you re-run your script, a new
graph should be created. However, for deployment the Arduino may reset
itself, running the `setup()` loop, and overwriting the data. The
`fileopt` `“extend”` does not overwrite the data, it instead “extends”
the existing data with the new data.

So for deployment, the scripts should look like:

```
void setup() {
  Serial.begin(9600);

  startEthernet();

  bool success;
  graph.maxpoints = 500;
  graph.fileopt = "extend"; // Remove this if you want the graph to be
overwritten
  success = graph.init();
  if(!success){while(true){}}
  graph.openStream();
}
```